ARTP-629
Moved chartIQ button
