+++ src/client/src/ui/spotTile/components/TileControls.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { PopoutIcon, usePlatform } from 'rt-components'
+import { PopoutIcon } from 'rt-components'
 import { styled } from 'rt-theme'
 
 export const TopRightButton = styled('button')`
@@ -16,37 +16,19 @@ export const TopRightButton = styled('button')`
   }
 `
 
-export const BottomRightButton = styled('button')`
-  position: absolute;
-  right: 0;
-  bottom: 0;
-  opacity: 0;
-  transition: opacity 0.2s;
-  padding: 0.25rem;
-`
-
 interface Props {
   canPopout?: boolean
   onPopoutClick?: () => void
-  displayCurrencyChart?: () => void
 }
 
-const TileControls: React.FC<Props> = ({ onPopoutClick, canPopout, displayCurrencyChart }) => {
-  const platform = usePlatform()
-  return (
-    <React.Fragment>
-      {canPopout && (
-        <TopRightButton onClick={onPopoutClick}>
-          <PopoutIcon width={0.8125} height={0.75} />
-        </TopRightButton>
-      )}
-      {platform.type !== 'browser' && (
-        <BottomRightButton onClick={displayCurrencyChart}>
-          <i className="fas fa-chart-bar" />
-        </BottomRightButton>
-      )}
-    </React.Fragment>
-  )
-}
+const TileControls: React.FC<Props> = ({ onPopoutClick, canPopout }) => (
+  <React.Fragment>
+    {canPopout && (
+      <TopRightButton onClick={onPopoutClick}>
+        <PopoutIcon width={0.8125} height={0.75} />
+      </TopRightButton>
+    )}
+  </React.Fragment>
+)
 
 export default TileControls
